refactor: refactored type checker and errors
chore: on ir just modifieid to use the reexports of hir
