+++ crates/ir/src/instructions.rs
@@ -1,6 +1,6 @@
 use common::Operator;
 use either::Either;
-use slynx_hir::model::{HirExpression, HirExpressionKind, HirStatement, HirStatementKind};
+use slynx_hir::{HirExpression, HirExpressionKind, HirStatement, HirStatementKind};
 
 use crate::{
     IRError, IRPointer, IRTypeId, Instruction, Label, Operand, Slot, SlynxIR, TempIRData, Value,
