+++ src/compilation_context/errors/hir.rs
@@ -16,25 +16,25 @@ impl SlynxContext {
                 expected_length,
                 received_length,
             } => {
-                let func_name = hir.get_name(*func_name);
+                let func_name = hir.get_name_from_pointer(*func_name);
                 format!(
                     "Function '{func_name}' expected to receive {expected_length} arguments, instead got {received_length} arguments"
                 )
             }
             HIRErrorKind::NotAFunction(name, ty) => {
-                let name = hir.get_name(*name);
+                let name = hir.get_name_from_pointer(*name);
                 format!(
                     "The value with name '{name}' is being used as a function, but its type is {ty:?}"
                 )
             }
             HIRErrorKind::NameNotRecognized(name) => {
-                let name = hir.get_name(*name);
+                let name = hir.get_name_from_pointer(*name);
                 format!(
                     "The name '{name}' is not recognized. Check if it exists or you wrote some typo"
                 )
             }
             HIRErrorKind::TypeNotRecognized(name) => {
-                let name = hir.get_name(*name);
+                let name = hir.get_name_from_pointer(*name);
                 format!("Type with name '{name}' is was not defined previously")
             }
             HIRErrorKind::InvalidFieldAccessTarget { ty } => {
@@ -50,24 +50,24 @@ impl SlynxContext {
             }
             HIRErrorKind::InvalidBinaryExpression { .. } => "Invalid binary expression".to_string(),
             HIRErrorKind::PropertyNotVisible { prop_name } => {
-                let prop_name = hir.get_name(*prop_name);
+                let prop_name = hir.get_name_from_pointer(*prop_name);
                 format!("Property with name '{prop_name}' is not visible")
             }
             HIRErrorKind::InvalidChild => {
                 "Invalid child. Component is not expecting children".to_string()
             }
             HIRErrorKind::InvalidType { ty, reason } => {
-                let ty = hir.get_name(*ty);
+                let ty = hir.get_name_from_pointer(*ty);
                 format!("Invalid type '{ty}' because it's {reason}")
             }
             HIRErrorKind::NameAlreadyDefined(name) => {
-                let name = hir.get_name(*name);
+                let name = hir.get_name_from_pointer(*name);
                 format!("The name '{name}' was already defined before. Use a different name")
             }
             HIRErrorKind::MissingProperty { prop_names } => {
                 let names = prop_names
                     .iter()
-                    .map(|v| hir.get_name(*v))
+                    .map(|v| hir.get_name_from_pointer(*v))
                     .collect::<Vec<&str>>()
                     .join(", ");
                 format!("Property(ies) named as {names} is required but wasn't provided")
@@ -75,21 +75,21 @@ impl SlynxContext {
             HIRErrorKind::PropertyNotRecognized { prop_names } => {
                 let names = prop_names
                     .iter()
-                    .map(|v| hir.get_name(*v))
+                    .map(|v| hir.get_name_from_pointer(*v))
                     .collect::<Vec<_>>()
                     .join(", ");
                 format!("Property(ies) named as {names} are not recognized for this object")
             }
             HIRErrorKind::RecursiveType { ty } => {
-                let ty = hir.get_name(*ty);
+                let ty = hir.get_name_from_pointer(*ty);
                 format!("The type named as '{ty}' is recursive at this point")
             }
             HIRErrorKind::InvalidStyleEvent { name } => {
-                let name = hir.get_name(*name);
+                let name = hir.get_name_from_pointer(*name);
                 format!("Invalid style event '{name}'")
             }
             HIRErrorKind::InvalidStyleDefinition { name } => {
-                let name = hir.get_name(*name);
+                let name = hir.get_name_from_pointer(*name);
                 format!("Invalid style definition '{name}'")
             }
         }
